fix unit test compilation; rename `DataBuilder` to `BytesBuilder`
