+++ Sources/ByteBuilder/ByteBuilder.swift
@@ -119,31 +119,6 @@ public struct ByteBuilder {
 }
 /*
 
-// MARK: Stream & Data Builder
-extension CompressionTechnique {
-    public struct DataBuilder {
-        public var data:[UInt8]
-        public var builder:ByteBuilder
-
-        public init(data: [UInt8] = [], builder: ByteBuilder = ByteBuilder()) {
-            self.data = data
-            self.builder = builder
-        }
-
-        public mutating func write(bit: Bool) {
-            if let wrote:UInt8 = builder.write(bit: bit) {
-                data.append(wrote)
-            }
-        }
-        public mutating func write(bits: [Bool]) {
-            builder.write(bits: bits, closure: { data.append($0) })
-        }
-        public mutating func finalize() {
-            builder.flush(into: &data)
-        }
-    }
-}
-
 // MARK: StreamBuilder
 extension CompressionTechnique {
     @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
