+++ Tests/SwiftCompressionTests/ByteBuilderTests.swift
@@ -1,12 +1,13 @@
 
 #if compiler(>=6.0)
 
+import ByteBuilder
 import Testing
 @testable import SwiftCompressionUtilities
 
-struct DataBuilderTests {
-    @Test func dataBuilder() {
-        var builder:CompressionTechnique.DataBuilder = .init()
+struct BytesBuilderTests {
+    @Test func bytesBuilder() {
+        var builder = BytesBuilder()
         builder.data.reserveCapacity(100)
         builder.write(bits: [true, false, true, false])
         builder.write(bits: [true, true, true, true, true])
